🤖 tests: share task service harness
_Generated with `xum` • Model: `openai:gpt-5.6-sol` • Thinking: `high` • Cost: `$`_

<!-- mux-attribution: model=openai:gpt-5.6-sol thinking=high costs= -->
